fixes to serialization and support for multi-dimensional arrays in serialization.

+++ errai-bus/src/main/java/org/jboss/errai/bus/server/io/JSONStreamDecoder.java
@@ -212,7 +212,7 @@ public class JSONStreamDecoder {
         long val = 0;
         double dVal = 0;
 
-        long factor = 1;
+        double factor = 1;
         boolean dbl = false;
 
         char[] buf = new char[21];
